Back-end/pyproject.toml
[project]
name = "fastapi-gemini-backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
